Let a client build without the daemon's Zig dependency
cmux-remote pulled in libghostty-vt unconditionally, but the only code
that uses it is the workspace service's server-side terminal model. A
client dialing a daemon needs none of that, and on iOS the dependency is
not merely dead weight: it makes a Rust toolchain and the iOS SDK
insufficient to build a client, because libghostty-vt is a Zig build.

The new default-on daemon-services feature carries the workspace
service, the handlers built on it, and the local bridge. MessageStream
and ServicesError move to a new message module, since both endpoints
need them and only their old home was daemon-side.

Building without the feature leaves the carrier, the authenticated
session, and WorkspaceClient intact, and needs no Zig at all.
