+++ cmux-tui/crates/cmux-remote/Cargo.toml
@@ -11,8 +11,11 @@ description = "Authenticated remote daemon, transports, and workspace services f
 workspace = true
 
 [features]
-default = ["iroh-transport"]
+default = ["iroh-transport", "daemon-services"]
 iroh-transport = ["dep:iroh", "dep:socket2"]
+# Host-side services. Off for client-only builds, which then need no Zig
+# toolchain because libghostty-vt is only used by the workspace service.
+daemon-services = ["dep:ghostty-vt"]
 
 [dependencies]
 anyhow.workspace = true
@@ -25,7 +28,7 @@ cmux-remote-protocol.workspace = true
 diffy.workspace = true
 futures-util.workspace = true
 getrandom.workspace = true
-ghostty-vt.workspace = true
+ghostty-vt = { workspace = true, optional = true }
 hmac.workspace = true
 http.workspace = true
 iroh = { workspace = true, optional = true }
