+++ cmux-tui/crates/cmux-remote/src/client.rs
@@ -9,7 +9,7 @@ use cmux_remote_protocol::{
 use tokio::sync::{Mutex, oneshot, watch};
 
 use crate::service::{ServiceMultiplexer, ServiceStream};
-use crate::services::MessageStream;
+use crate::message::MessageStream;
 
 type PendingResponse = Result<RpcResponse, String>;
 type PendingRequests = Arc<Mutex<HashMap<RequestId, oneshot::Sender<PendingResponse>>>>;
